refactor --start-time flags to _cli_utils
